Fiddle with site colours
I've redefined the colour variables so they have a semantic meaning. There's an
accent colour, selected colour and highlight colour alongside three text colours.

I've modified the style of the site a bit so there's less colour overall. The
links on the homepage are now styled the same as text. The colour used for links
is now the same as the footer and header colour.